css: compute the D50 reference white in f32 so achromatic colors convert to lab(l 0 0)
The D50 white point used by the xyz-d50 <-> lab conversions was computed
in f64 and cast to f32. Those constants are one f32 ulp away (in x and z)
from the values lightningcss uses, which computes them in f32, and from
where the f32 D65 -> D50 adaptation matrix puts achromatic colors. The
division by the reference white therefore left x, y and z an ulp apart
for any grey, and 500 * (f0 - f1) / 200 * (f1 - f2) turned that into
a/b residuals around 1e-5: lab(from gray l a b) printed
lab(53.585% -.0000298023 .0000119209) and lch(from gray l c h) printed
lch(53.585% .0000320981 158.199), where lightningcss prints 0 0.

Computing the constants in f32 makes every lab()/lch() result checked
against lightningcss 1.30.2 byte-identical.

+++ src/css/values/color.rs
@@ -2726,11 +2726,10 @@ fn polar_to_rectangular(l: f32, c: f32, h: f32) -> (f32, f32, f32) {
     (l, a, b)
 }
 
-const D50: [f32; 3] = [
-    (0.3457f64 / 0.3585f64) as f32,
-    1.00000,
-    ((1.0f64 - 0.3457f64 - 0.3585f64) / 0.3585f64) as f32,
-];
+// Deliberately computed in f32 (as lightningcss does), not in f64 and then cast: the f32 values
+// nearest to the exact quotients are one ulp away from where the f32 XYZd65 -> XYZd50 matrix below
+// puts achromatic colors, which leaves lab(from gray l a b) with a, b around 1e-5 instead of 0 0.
+const D50: [f32; 3] = [0.3457 / 0.3585, 1.00000, (1.0 - 0.3457 - 0.3585) / 0.3585];
 
 // ──────────────────────────────────────────────────────────────────────────
 // Handwritten conversions.
